+++ apps/client/themes/default/commons/site-logo.tsx
