+++ apps/client/themes/list/commons/publishedAt.tsx
@@ -1,8 +1,11 @@
-import { IconBook } from '../../../components/icons';
-import formatDate from '../../../lib/utils/formatDate';
+import classNames from 'classnames';
+
+import formatDate from '@/lib/utils/formatDate';
+
+import { IconBook } from '@/components/icons';
 
 export const PublishedAt = ({ publishedAt, className, reading_time }) => (
-  <dl className={className + ' ' + 'text-gray-500 dark:text-gray-400'}>
+  <dl className={classNames('text-gray-500 dark:text-gray-400', className)}>
     <dt className="sr-only">Published on</dt>
     <dd className=" font-medium leading-6 ">
       <time dateTime={publishedAt}>{formatDate(publishedAt)}</time>
